git-svn-id: http://svn.code.sf.net/p/emuscriptoria/code@294 88e056c0-0935-4f5d-8a44-a8e7373e37e9

+++ compila.php
@@ -540,16 +540,9 @@ function compg($out){
   ob_start();$x='JAs';$y=0x2000;$title='jupiler';
   require'emu_ace.php';
   compg('JAs');
-//
-/*
-  unlink('z80elite.js');
-  unlink('z80.js');
-  unlink('z80m.js');
-  unlink('z80p.js');
-  unlink('z80mp.js');
-  unlink('z80pc.js');
-  unlink('z80mpc.js');
 //*/
-  unlink('aa.rom');
-  unlink('temp.zip');
+  if( $slow ){
+    unlink('aa.rom');
+    unlink('temp.zip');
+  }
 ?>
